Rename usersReducer by chatReducer for better understanding
