+++ mobile/reducers/chatReducer.js
